+++ routes/jwt.mjs
@@ -1,4 +1,4 @@
-import Router from 'koa-router';
+import Router from '@koa/router';
 import jwt from 'koa-jwt';
 import config from '../services/config';
 import User, { ExpiredPasswordError } from '../models/user';