+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Export metrics for prometheus
 ### Fix
 * bump validator to 12.1.0 & deleted node-input-validator
+* Updated csv flatten middleware to common csv formatter for rides
 * Deleted password from token
 ## Version 1.4.4
 ### Features